SHIT COMPILES
feat(spi): update message converter to use encoder data for odometry and joint state calculations

+++ star-ros2/src/star_spi_bridge/src/spi_driver.cpp
@@ -180,7 +180,7 @@ bool SpiDriver::decode_frame(
 
   // Validate Frame Size
   // Header(8) + Payload(len) + CRC(4)
-  if (frame.size() != 8 + len + 4) {
+  if (frame.size() != static_cast<size_t>(8 + len + 4)) {
     return false;
   }
